code
thêm 3 thư viện để icon

Lỗi bạn đang gặp có thể xuất phát từ việc bạn chưa import và sử dụng thư viện FontAwesome một cách đúng cách. Để sử dụng FontAwesome trong ứng dụng của bạn, bạn cần cài đặt và import thư viện này. Hãy đảm bảo bạn đã thực hiện các bước sau:

Cài đặt FontAwesome bằng npm hoặc yarn:
bash
Copy code
npm install --save @fortawesome/fontawesome-svg-core
npm install --save @fortawesome/free-solid-svg-icons
npm install --save @fortawesome/react-fontawesome
thay đổi thư mục chưa profile
